Flatten folder structure plus some minor refactorings
