+++ src/Azure.Deployments.Extensibility.Core/V2/Extensions/IServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Microsoft Corporation.
+// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
 using Microsoft.AspNetCore.Http.Json;
