+++ src/Azure.Deployments.Extensibility.Core/V2/Validators/IResourceConfigValidator.cs
@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using Azure.Deployments.Extensibility.Core.V2.Models;
 using System.Text.Json.Nodes;
 
-namespace Azure.Deployments.Extensibility.Core.V2.Models.Validation
+namespace Azure.Deployments.Extensibility.Core.V2.Validators
 {
     public interface IResourceConfigValidator : IValidator<JsonObject?, IReadOnlyList<ErrorDetail>>
     {
