+++ src/Azure.Deployments.Extensibility.Core/V2/Validators/IResourceTypeValidator.cs
@@ -1,7 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Azure.Deployments.Extensibility.Core.V2.Models.Validation
+using Azure.Deployments.Extensibility.Core.V2.Models;
+
+namespace Azure.Deployments.Extensibility.Core.V2.Validators
 {
     public interface IResourceTypeValidator : IValidator<string, IReadOnlyList<ErrorDetail>>
     {
