+++ src/Azure.Deployments.Extensibility.Core/V2/Validators/IValidator.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Azure.Deployments.Extensibility.Core.V2.Models.Validation
+namespace Azure.Deployments.Extensibility.Core.V2.Validators
 {
     public interface IValidator<TValue, out TResult>
     {
