+++ src/Azure.Deployments.Extensibility.Core/V2/Validators/PassthroughResourceConfigValidator.cs
@@ -1,11 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using Azure.Deployments.Extensibility.Core.V2.Models;
 using System.Text.Json.Nodes;
 
-namespace Azure.Deployments.Extensibility.Core.V2.Models.Validation
+namespace Azure.Deployments.Extensibility.Core.V2.Validators
 {
-    public class PassthroughResourceConfigValidator : IResourceConfigValidator
+    public sealed class PassthroughResourceConfigValidator : IResourceConfigValidator
     {
         private static readonly Lazy<PassthroughResourceConfigValidator> LazyInstance = new(() => new());
