+++ src/Azure.Deployments.Extensibility.Core/V2/Validators/ResourceConfigSchemaValidator.cs
@@ -1,32 +1,25 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using Azure.Deployments.Extensibility.Core.V2.Models;
 using Azure.Deployments.Extensibility.Core.V2.Utils;
 using Json.Pointer;
 using Json.Schema;
 using System.Text.Json.Nodes;
 
-namespace Azure.Deployments.Extensibility.Core.V2.Models.Validation
+namespace Azure.Deployments.Extensibility.Core.V2.Validators
 {
-    public class ResourceConfigSchemaValidator : IResourceConfigValidator
+    public class ResourceConfigSchemaValidator(JsonSchema configSchema, bool configRequired = true) : IResourceConfigValidator
     {
         private readonly static JsonPointer BaseJsonPointer = JsonPointerBuilder.Build<ResourceRequestBody>(x => x.Config!);
 
-        private readonly JsonSchemaValidator validator;
+        private readonly JsonSchemaValidator validator = new(configSchema);
 
-        private readonly bool configRequired;
-
-        public ResourceConfigSchemaValidator(JsonSchema configSchema, bool configRequired = true)
-        {
-            this.validator = new JsonSchemaValidator(configSchema);
-            this.configRequired = configRequired;
-        }
-
-        public IReadOnlyList<ErrorDetail> Validate(JsonObject? config)
+        public virtual IReadOnlyList<ErrorDetail> Validate(JsonObject? config)
         {
             if (config is null)
             {
-                if (this.configRequired)
+                if (configRequired)
                 {
                     return new[] { new ErrorDetail("InvalidConfig", "Config cannot be null or undefined.", BaseJsonPointer) };
                 }
